// packages/js/data/src/products/action-types.ts
export enum TYPES {
	GET_PRODUCT_SUCCESS = 'GET_PRODUCT_SUCCESS',
	GET_PRODUCT_ERROR = 'GET_PRODUCT_ERROR',
	GET_PRODUCTS_SUCCESS = 'GET_PRODUCTS_SUCCESS',
	GET_PRODUCTS_ERROR = 'GET_PRODUCTS_ERROR',
	GET_PRODUCTS_TOTAL_COUNT_SUCCESS = 'GET_PRODUCTS_TOTAL_COUNT_SUCCESS',
	GET_PRODUCTS_TOTAL_COUNT_ERROR = 'GET_PRODUCTS_TOTAL_COUNT_ERROR',
	SET_IS_REQUESTING = 'SET_IS_REQUESTING',
	UPDATE_PRODUCT = 'UPDATE_PRODUCT',
	UPDATE_PRODUCT_ERROR = 'UPDATE_PRODUCT_ERROR',
}

export default TYPES;